reorganized folder and made regions a provider property
