--- src/constants.ts
@@ -1,12 +1,0 @@
-export const AVG_CPU = 'avgCpu';
-export const MAX_CPU = 'maxCpu';
-export const AVG_MEMORY = 'avgMemory';
-export const MAX_MEMORY = 'maxMemory';
-export const DISK_READ_OPS = 'diskReadOps';
-export const DISK_WRITE_OPS = 'diskWriteOps';
-export const MAX_NETWORK_BYTES_IN = 'maxNetworkBytesIn';
-export const MAX_NETWORK_BYTES_OUT = 'maxNetworkBytesOut';
-export const AVG_NETWORK_BYTES_IN = 'avgNetworkBytesIn';
-export const AVG_NETWORK_BYTES_OUT = 'avgNetworkBytesOut';
-export const ALB_REQUEST_COUNT = 'albRequestCount';
-export const APIG_REQUEST_COUNT = 'apigRequestCount';
