+++ src/index.ts
@@ -1,4 +1,4 @@
-import { AwsUtilization } from './aws-utilization.js';
+import { AwsUtilization } from './widgets/aws-utilization.js';
 import { AwsUtilizationProvider } from './aws-utilization-provider.js';
 import { AwsUtilizationRecommendations } from './widgets/aws-utilization-recommendations.js';
 export {
