+++ src/widgets/utilization-recommendations-ui/confirm-recommendations.tsx
@@ -5,7 +5,7 @@ import {
 } from '@chakra-ui/react';
 import { ArrowBackIcon } from '@chakra-ui/icons';
 import { ConfirmSingleRecommendation } from './confirm-single-recommendation.js';
-import { ConfirmRecommendationsProps } from '../utilization-recommendations-types.js';
+import { ConfirmRecommendationsProps } from '../../types/utilization-recommendations-types.js';
 import { actionTypeText } from '../../types/types.js';
 import { filterUtilizationForActionType } from '../../utils/utilization.js';
