+++ src/widgets/utilization-recommendations-ui/recommendations-table.tsx
@@ -11,7 +11,7 @@ import isEmpty from 'lodash.isempty';
 import ServiceTableRow from './service-table-row.js';
 import { Utilization, actionTypeText, ActionType } from '../../types/types.js';
 import { filterUtilizationForActionType, sentenceCase } from '../../utils/utilization.js';
-import { RecommendationsTableProps } from '../utilization-recommendations-types.js';
+import { RecommendationsTableProps } from '../../types/utilization-recommendations-types.js';
 import { Drawer, DrawerOverlay, DrawerContent, DrawerCloseButton, DrawerHeader, DrawerBody } from '@chakra-ui/react';
 import { ChevronDownIcon, InfoIcon, ExternalLinkIcon } from '@chakra-ui/icons';
